requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
python-decouple>=3.8
Pillow>=10.0.0
django-filter>=23.5
